added queue controller class, queue processor: queue members are saved in blocking Queue. logging management, exception handling

+++ src/main/java/com/queuesystem/percistance/model/QueueMember.java
@@ -15,7 +15,7 @@ public class QueueMember {
 
 	private String name;
 
-	private int queueNumber;
+	private long queueNumber;
 
 	@Enumerated(EnumType.STRING)
 	private QueueMemberStatus status;
